+++ CHANGELOG.md
@@ -2,4 +2,5 @@
 
 ## v1.0.0
 
-Initial Release
+- Initial release based on Grafana 7.2.0.
+- Includes Redis CLI Panel.